Do not crash if matplotlib Gtk backend fails
Ignore the RuntimeError exception of the Gtk matplotlib backend if
an X11 display could not be opened. Import scape without plotting
functionality instead. This is useful because the Gtk backend tends
to be the default one on Linux, and Linux is frequently used for
remote testing and debugging of scape, without X11 forwarding too.