Use the new katdal weather interface
This is a unified way to access weather / enviro sensors via properties on
the katdal.DataSet object. It is no longer necessary for scape to track the
details of the underlying file formats. On the other hand, this serves up
the interpolated sensor data on the data timestamps instead of the original
raw sensor events.