Allow scape to load v4 RDB files
This adds a check for filenames (or more generally, URLs) with a .rdb
extension and routes them to the 'new' loader based on katdal
(misleadingly called `hdf5_load`...). Previously they defaulted to the
'old' HDF5 loader, which then croaked.

Also avoid calling :func:`katdal.get_ants` on the filename, which is
not implemented for v4 datasets. Instead, open the dataset first using
the suggested antenna A and then continue as usual.

This addresses JIRA ticket MRTS-166.