Load the appropriate az/el sensors for v4 files
The names have changed in v4, so load them with their raw names, which
still respects the selected pointing. An alternative would have been
to map them to `Antennas/mxxx/name` but that needs changes to katdal
as well (is it worth it?). Use the opportunity to clean up the naming
logic.

This addresses JIRA tickets MRTS-167 and MKAIV-1221.