SPR1-40 Use print-like handler if no log config
Instead of running logging.basicConfig inside a library (a no-no)
do a very basicConfig-like configuration at the module level. This
logs DEBUG and above to sys.stderr using the basic format
(not unlike a plain old print statement), thereby emulating
the expected user output but without triggering the very verbose
debug output of e.g. Numba and Matplotlib.

While this behaviour is strongly discouraged by the logging
documentation, the saving grace is that the configured handler
also has a filter that disables the handler once the user properly
configures logging at the application or script level.

This copies the behaviour of katdal introduced in ska-sa/katdal#6.